requirements.txt
opencv-python>=4.8.0
pillow>=10.0.0
requests>=2.31.0
psutil>=5.9.0
easyocr>=1.7.0
ultralytics>=8.0.0
numpy>=1.24.0
pyyaml>=6.0